chore: move reference rendering to a separate file
